detector/requirements.txt
# DO NOT INSTALL ON HOST - THIS WILL BE RUN IN CONTAINER

gunicorn
matplotlib>=3.2.2
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
tqdm>=4.41.0
pandas>=1.1.4
seaborn>=0.11.0
protobuf<=3.20.1
ipython
flask